ClaudeSafe/requirements.txt
anthropic>=0.39.0
python-telegram-bot>=21.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
